server/requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
bcrypt==3.1.7
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.5
django-background-tasks==1.2.0
django-compat==1.0.15
django-filter==2.1.0
django-jet==1.0.8
djangorestframework==3.9.3
fabric==2.5.0
feedparser==5.2.1
firebase-admin==2.18.0
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.19.0
google-resumable-media==0.3.3
googleapis-common-protos==1.6.0
grpcio==1.23.0
httplib2==0.13.1
idna==2.8
invoke==1.3.0
isort==4.3.21
lazy-object-proxy==1.4.2
Markdown==3.1
mccabe==0.6.1
migrate==0.3.8
msgpack==0.6.1
mysqlclient==1.4.2.post1
paramiko==2.6.0
pbr==5.4.2
Pillow==6.1.0
protobuf==3.9.1
psycopg2==2.8.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pylint==2.3.1
PyNaCl==1.3.0
pytz==2019.1
requests==2.22.0
rsa==4.0
sentry-sdk==0.11.2
six==1.12.0
sqlparse==0.3.0
stevedore==1.30.1
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
uWSGI==2.0.18
virtualenv==16.7.5
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wrapt==1.11.2
